Implement field value persistence
This commit introduces a `%.values` hash to the `Rakuwa::Form` class.
This hash is used to store and retrieve field values, allowing for
persistence of form data across requests if it's not present in the
query parameters.

The `submit` method has also been updated to prevent duplicate submit
button names in the `@.submits-names` array.